Extract all webhook REST-type stuff into a new class
